# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "xinzhi-daoxue-api",
    "description": "芯智导学阶段 2.2 API。统一 Agent 注册、快速路由、受控降级与多星辰工作流共用同一 Provider 调用链。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/knowledge/search": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Search Knowledge",
        "operationId": "search_knowledge_api_v1_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/evaluate-query": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Evaluate Query",
        "operationId": "evaluate_query_api_v1_knowledge_evaluate_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/rag-search": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Rag Search",
        "operationId": "rag_search_api_v1_knowledge_rag_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/images/{course_id}/{relative_path}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Knowledge Image",
        "operationId": "knowledge_image_api_v1_knowledge_images__course_id___relative_path__get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "relative_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Relative Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/documents/{course_id}/{relative_path}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Knowledge Document",
        "operationId": "knowledge_document_api_v1_knowledge_documents__course_id___relative_path__get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "relative_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Relative Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeCourseId": {
        "type": "string",
        "enum": [
          "CT",
          "AE",
          "DE"
        ],
        "title": "KnowledgeCourseId"
      },
      "KnowledgeHit": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id",
            "default": ""
          },
          "evidence_id": {
            "type": "string",
            "title": "Evidence Id",
            "default": ""
          },
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "default": ""
          },
          "course_id": {
            "$ref": "#/components/schemas/KnowledgeCourseId"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "chapter": {
            "type": "string",
            "title": "Chapter",
            "default": ""
          },
          "section": {
            "type": "string",
            "title": "Section",
            "default": ""
          },
          "document_path": {
            "type": "string",
            "title": "Document Path"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "unknown"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "score": {
            "type": "number",
            "minimum": 0.0,
            "title": "Score"
          },
          "score_components": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Score Components"
          },
          "source_ref": {
            "type": "string",
            "title": "Source Ref"
          },
          "document_checksum": {
            "type": "string",
            "title": "Document Checksum",
            "default": ""
          },
          "related_images": {
            "items": {
              "$ref": "#/components/schemas/RelatedImage"
            },
            "type": "array",
            "title": "Related Images"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "course_id",
          "course_name",
          "document_path",
          "title",
          "content",
          "score",
          "source_ref"
        ],
        "title": "KnowledgeHit"
      },
      "KnowledgeSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query"
          },
          "course_ids": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeCourseId"
            },
            "type": "array",
            "title": "Course Ids"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "KnowledgeSearchRequest"
      },
      "KnowledgeSearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeHit"
            },
            "type": "array",
            "title": "Hits"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeSourceStatus"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "KnowledgeSearchResponse"
      },
      "KnowledgeSourceStatus": {
        "properties": {
          "course_id": {
            "$ref": "#/components/schemas/KnowledgeCourseId"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "document_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Document Count"
          },
          "chunk_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Chunk Count"
          },
          "indexed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "course_id",
          "course_name",
          "available",
          "document_count",
          "chunk_count"
        ],
        "title": "KnowledgeSourceStatus"
      },
...
      "RAGSearchRequest": {
        "properties": {
          "query_text": {
            "type": "string",
            "maxLength": 1000,
            "title": "Query Text",
            "default": ""
          },
          "course_id": {
            "$ref": "#/components/schemas/KnowledgeCourseId"
          },
          "image_resource_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Resource Uri"
          },
          "intent": {
            "type": "string",
            "title": "Intent",
            "default": "general_qa"
          },
          "target_agent_id": {
            "type": "string",
            "title": "Target Agent Id",
            "default": "LEARN_01_KNOWLEDGE_QA_V1"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "content_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Types"
          },
          "include_images": {
            "type": "boolean",
            "title": "Include Images",
            "default": true
          },
          "use_reranker": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Reranker"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "course_id"
        ],
        "title": "RAGSearchRequest"
      },
      "RelatedImage": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "resource_uri": {
            "type": "string",
            "title": "Resource Uri"
          },
          "caption": {
            "type": "string",
            "title": "Caption",
            "default": ""
          },
          "description_source": {
            "type": "string",
            "title": "Description Source",
            "default": "source_text"
          },
          "course_id": {
            "type": "string",
            "title": "Course Id",
            "default": ""
          },
          "parent_document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Document Id"
          },
          "parent_chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Chunk Id"
          },
          "image_type": {
            "type": "string",
            "title": "Image Type",
            "default": "unknown"
          },
          "score": {
            "type": "number",
            "minimum": 0.0,
            "title": "Score",
            "default": 0
          },
          "retrieval_channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Retrieval Channels"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "image_id",
          "resource_uri"
        ],
        "title": "RelatedImage"
      },
      "RetrievalResult": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "normalized_query": {
            "type": "string",
            "title": "Normalized Query"
          },
          "course_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Course Ids"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeHit"
            },
            "type": "array",
            "title": "Hits"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "retrieval_mode": {
            "type": "string",
            "title": "Retrieval Mode",
            "default": "local_lexical_v2"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "latency_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Latency Ms"
          },
          "image_hits": {
            "items": {
              "$ref": "#/components/schemas/RelatedImage"
            },
            "type": "array",
            "title": "Image Hits"
          },
          "rag_status": {
            "type": "string",
            "title": "Rag Status",
            "default": "disabled"
          },
          "embedding_status": {
            "type": "string",
            "title": "Embedding Status",
            "default": "disabled"
          },
          "vector_store_status": {
            "type": "string",
            "title": "Vector Store Status",
            "default": "disabled"
          },
          "reranker_status": {
            "type": "string",
            "title": "Reranker Status",
            "default": "disabled"
          },
          "query_modalities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Query Modalities"
          },
          "retrieval_trace_id": {
            "type": "string",
            "title": "Retrieval Trace Id",
            "default": ""
          },
          "index_version": {
            "type": "string",
            "title": "Index Version",
            "default": ""
          },
          "trace": {
            "additionalProperties": true,
            "type": "object",
            "title": "Trace"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query",
          "normalized_query",
          "course_ids",
          "latency_ms"
        ],
        "title": "RetrievalResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}